+++ src/main/java/insane96mcp/progressivebosses/ProgressiveBosses.java
@@ -4,6 +4,7 @@ import insane96mcp.progressivebosses.commands.PBCommand;
 import insane96mcp.progressivebosses.module.dragon.data.DragonStatsReloadListener;
 import insane96mcp.progressivebosses.module.dragon.entity.Larva;
 import insane96mcp.progressivebosses.module.dragon.phase.CrystalRespawnPhase;
+import insane96mcp.progressivebosses.module.dragon.phase.DragonStrafePillarPhase;
 import insane96mcp.progressivebosses.module.dragon.phase.PBDragonStrafePlayerPhase;
 import insane96mcp.progressivebosses.module.elderguardian.data.ElderGuardianStatsReloadListener;
 import insane96mcp.progressivebosses.module.wither.data.WitherStatsReloadListener;
@@ -53,6 +54,7 @@ public class ProgressiveBosses {
 
 		CrystalRespawnPhase.init();
 		PBDragonStrafePlayerPhase.init();
+		DragonStrafePillarPhase.init();
 
 		DispenserBlock.registerBehavior(Items.WITHER_SKELETON_SKULL, new WitherSkullDispenseBehavior());
 	}