+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 * Remove Blinding Shulker Bullets
-* Add AoE effects (blindness? levitation?) on caged towers
+* change on break crystal to strafe pillar
 
 ## Upcoming
 * Data packable Ender Dragon
@@ -15,6 +15,7 @@
   * Crystals attached to the dragon when destroyed now deal 20% of current health as damage
   * Removed blindness minions. Any shulker now has a chance to fire a blinding bullet
   * Minions now shoot 4 bullets before going into cooldown
+  * Minions now spawn on towers
   * Removed ender dragon acid ball shotgun, now shoots them in rapid succession
   * Ender Dragon now stays much less time at the center podium
   * Respawning crystals now also summons some phantoms